Add files via upload
Here is an assignment for Internshala Pi RSquare . I created front-end only. Here intern3.html is a main program code. when click this page it will go and execute the all programs and imagepreview.html, videopreview.html these pages are sub programs. I developed the merging audio and Image as an video.